Extract the placeholder avatar image generator from the NSE
